(PC-34315) after review refacto log send
