+++ src/features/offer/components/OfferContent/ChronicleSection/ChronicleSection.web.tsx
@@ -16,8 +16,7 @@ import { ChronicleSectionProps } from './types'
 
 export const ChronicleSection = (props: ChronicleSectionProps) => {
   const { isDesktopViewport } = useTheme()
-  const { data, title, subtitle, ctaLabel, navigateTo, style, onSeeMoreButtonPress, fromOfferId } =
-    props
+  const { data, title, subtitle, ctaLabel, navigateTo, style, onSeeMoreButtonPress } = props
 
   return isDesktopViewport ? (
     <View style={style}>
@@ -36,11 +35,7 @@ export const ChronicleSection = (props: ChronicleSectionProps) => {
         </Row>
         {subtitle ? <StyledSubtitle>{subtitle}</StyledSubtitle> : null}
       </Gutter>
-      <StyledChronicleCardlist
-        data={data}
-        onSeeMoreButtonPress={onSeeMoreButtonPress}
-        fromOfferId={fromOfferId}
-      />
+      <StyledChronicleCardlist data={data} onSeeMoreButtonPress={onSeeMoreButtonPress} />
     </View>
   ) : (
     <ChronicleSectionBase {...props} />
