+++ src/features/offer/components/OfferContent/ChronicleSection/types.ts
@@ -11,5 +11,4 @@ export type ChronicleSectionProps = {
   navigateTo: InternalNavigationProps['navigateTo']
   onSeeMoreButtonPress?: (chronicleId: number) => void
   style?: StyleProp<ViewStyle>
-  fromOfferId?: number
 }
